Switch folder; dependency injection; note about moving to extensions
