+++ src/appCommands/viewer.js
@@ -1,20 +1,6 @@
 import { redux } from 'ohif-core';
-import store from './store';
+import store from './../store';
 const { setViewportActive } = redux.actions;
-
-const definitions = {
-  // Next/Previous active viewport
-  incrementActiveViewport: {
-    commandFn: actions.updateActiveViewport,
-    storeContexts: ['viewports'],
-    options: { direction: 1 },
-  },
-  decrementActiveViewport: {
-    commandFn: actions.updateActiveViewport,
-    storeContexts: ['viewports'],
-    options: { direction: -1 },
-  },
-};
 
 const actions = {
   updateViewportDisplaySet: ({ direction }) => {
@@ -30,6 +16,20 @@ const actions = {
     newIndex = newIndex < 0 ? maxIndex : newIndex;
 
     store.dispatch(setViewportActive(newIndex));
+  },
+};
+
+const definitions = {
+  // Next/Previous active viewport
+  incrementActiveViewport: {
+    commandFn: actions.updateActiveViewport,
+    storeContexts: ['viewports'],
+    options: { direction: 1 },
+  },
+  decrementActiveViewport: {
+    commandFn: actions.updateActiveViewport,
+    storeContexts: ['viewports'],
+    options: { direction: -1 },
   },
 };
